+++ src/main/java/cart/dao/ProductDao.java
@@ -1,6 +1,6 @@
 package cart.dao;
 
-import cart.domain.Product;
+import cart.domain.product.Product;
 
 import java.util.List;
