+++ src/main/java/cart/domain/product/ProductPrice.java
@@ -1,4 +1,4 @@
-package cart.domain;
+package cart.domain.product;
 
 import java.util.Objects;
