+++ src/test/java/cart/controller/ProductControllerTest.java
@@ -1,7 +1,7 @@
 package cart.controller;
 
 import cart.dao.ProductDao;
-import cart.domain.Product;
+import cart.domain.product.Product;
 import cart.dto.ProductRequest;
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.AfterEach;
