+++ src/test/java/cart/domain/ProductImageTest.java
@@ -1,5 +1,6 @@
 package cart.domain;
 
+import cart.domain.product.ProductImage;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
