+++ src/test/java/cart/service/ProductServiceTest.java
@@ -1,6 +1,6 @@
 package cart.service;
 
-import cart.domain.Product;
+import cart.domain.product.Product;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
